Add reload method and /reload console command.
Attempts to reload the module instead of re-executing the process. It's
less reliable, but it's also much faster and much smoother. No need to
disconnect/reconnect. Works most of the time, but there might be subtle
bugs still... Sometimes Bambot stops responding to certain things and it
isn't clear if it's related to this or something else. If the module file
has compilation errors then it will pretty much cause a crash and require
restarting... I tried to detect that with 'eval "require ..."', but it
isn't working. So it's best to check (e.g., :!perl -c %) before attempting
to reload.