Strip CR (\r) characters from lines...
I haven't taken the time to formally re-read "the" RFC, but it sounds like
IRC messages are supposed to be CRLF terminated (whoops). The LF should
already be removed with chomp, but I guess I have to strip CR manually. I
think that it should be desirable to strip all instances out. I was seeing
some strange output on the console that I attributed to CRs, though I'm
not sure if they were actually the culprits. Nevertheless, they can GTFO.