Adding a rudimentary \o/ handler.
Only trusted AKA "friendly" nicks are currently supported. That should be
changed to friendly "users" instead of nicks and the number of automatic
responses per minute should be throttled to make sure the bot doesn't spam
a channel... For now I will trust my friendly list to not abuse him (and
trust the ops to hold *them* responsible if they do).