h4x: Make tilde optional in the ident part... xD
I've still been too lazy to refresh myself on the IRC protocol, but I
noticed a user logged in via freenode's Web interface and there was no
tilde in their "identity": nick!~user@host. This allows them to still be
processed without Perl errors, but that doesn't make it necessarily
correct or robust. :-X I don't know yet what it means if the tilde is or
isn't there. :P