Silly feature: says your name if you ask.
At some point somebody in channel said "say my name, say my name" to
somebody (and/or anybody). Perhaps even to Bambot himself. It was a
while ago (I'm sorting through a bunch of uncommitted changes). I
thought it would be funny "teach" him to respond to it and this is the
silly little piece that came of it.