Add .copyright-files script for apply-copyright.
Should have been in last commit. Used to list the files that need the
copyright applied/updated. It's a bit of a hack since I didn't want to
mess with 'finding' the files given how few they are and how specific
their rules are. In practice any file in bin/ and any .pm in lib/ should
be good too. :)