Separate version into a new .pm.
I discovered that the build helpers don't work with an incomplete perl
environment because 'require q{lib/Bambot}' would fail since its
dependencies weren't installed. This made the version getting fail,
which caused the build scripts to fail. This is more or less a contrived
hack to move the version into a separate Perl module so it can be loaded
separately by these build helpers. The main module then uses it too.

It seems to work, though I doubt that this is the best way to do this...
>:) I wonder how ExtUtils::MakeMaker parses the version, since it seems
to work even when the dependencies aren't installed. Meh.