Bugfix: Fix targetting pattern.
Targetting needs to ignore white-space in the nick[1] and for now will
require that white-space follows the colon.

[1] Since Bambot doesn't currently keep track of users in a channel
(perhaps this should be the next feature) and therefore can't match
against a list of known users he can only guess what is a nick and what
isn't. This might be for the best anyway since a typo might otherwise
result in once again multiple instances handling a command[2].

[2] Perhaps instead Bambot's could detect each other in channel and
decide upon which one is "dominant". :) I'm not sure how to go about
that in a netiquette-friendly way. Perhaps CTCP VERSION?

Being anchored to the beginning of the message also becomes kind of
significant. >:)