Reformat: Convert to same-line-braces[1].
[1] Experimentally...

I normally prefer to have braces on their own lines (so they line up
horizontally), but I've kind of been giving it a try with them on the
same line as the respective statement since that is apparently the
preferred style by the Perl community (they also apparently do it in the
Linux kernel so it might not hurt to get used to it). I'm on the fence
about it right now. I don't like it, but I don't mind it, and bleh. So
this commit changes Bambot into it for now. It should make the code a
bit more vertically terse (for better or worse), and I guess I'll have
to see how I find maintaining it like this. Worst case I should be able
to reverse patch this to go back. >:(