Reformat: Normalize control statements spacing.
I generally hate to see control statements (particularly ones that
require parenthesis for their conditions, etc.) having a space between
the statement keyword and conditions/whatever. The way that for works in
Perl have sort of loosened me up on that and I typically end up doing a
for with a space whether or not there's an explicit loop variable or
not, but I still generally avoid doing this with other statements that
don't support a variable between the keyboard and the condition.
Evidently I've mistakenly done a that few times though so here I am
cleaning it up. >:(