h4x: Periodically test health of connection.
This is my hack of a solution to bambot being disconnected by networking
hiccups. There doesn't appear for there to be any way to know if the
socket is healthy without using it. For now, a CTCP PING to the master,
but if that becomes annoying it will need to be changed. Ideally to
something that the server is going to know what to do with, and ideally
that what to do is nothing.