Move log file compression into CLI.
I'm not sure if these are being triggered during run-time, but bambot
seems to get himself hung when log files are compressed at some point.
In the future we can maybe fork a process to do the compression so that
logging isn't blocked, but in the meantime it's enough to remove
compression entirely until the user (myself) invokes it.