h4x: Catch unhandled exceptions in read loop.
In a desperate attempt to prevent Bambot from crashing lets catch
exceptions and log them and then continue on. Hopefully if there are
serious bugs the logs will be detailed enough to piece it apart, and if
there are subtle bugs he'll just live on and continue to work, skipping
the job that failed.