Always log verbose to storage with filterable tag.
By logging all messages to storage, even if the message is verbose and
verbose isn't enabled, it allows us to peer into much more detailed logs
in the event of a bug in "production". The new --log viewer mode can be
used to filter verbose messages out while reading the logs.
Alternatively, something as simple as `grep -v \|VERBOSE' would probably
suffice in a pinch.