Add recursive solution to reload run method.
The run method was essentially the one thing that we couldn't hot swap
and that's because it's the main method: you never exit it until Bambot
is killed. To solve the problem (without jumping through hoops or
dirtying the caller, and without disconnecting/reconnecting, etc.), I've
decided to die() with a specific exception which the run method can
catch and handle special. When caught it will log the result to the
appropriate place for the user and then recurse into itself.